luci-base: form.js: drop usage of option.template
The .template property of options was never used in form.js but used as a
legacy hack for semi automatic Lua to JS map conversions.

Drop its usage from form.js to avoid accidentally calling replace() on
non-string values in case an option happens to declare a private "template"
property.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>